// src/sdk/component-definition-types.ts
export interface TransformContext<
  TConfig extends Record<string, unknown> = Record<string, unknown>,
  TInputs extends Record<string, unknown> = Record<string, unknown>,
> {
  config: TConfig;
  inputs: TInputs;
}
// ...
export interface StaticComponentTransformContext<
  TConfig extends Record<string, unknown> = Record<string, unknown>,
  TInputs extends Record<string, unknown> = Record<string, unknown>,
  TFixtureData = unknown,
> {
  config: TConfig;
  inputs: TInputs;
  fixtureData: TFixtureData;
}

export interface StaticComponentTransformModule<
  TConfig extends Record<string, unknown> = Record<string, unknown>,
  TInputs extends Record<string, unknown> = Record<string, unknown>,
  TFixtureData = unknown,
  TOutputs extends Record<string, unknown> = Record<string, unknown>,
> {
  outputSchema: ZodTypeAny;
  transform: (
    context: StaticComponentTransformContext<TConfig, TInputs, TFixtureData>,
  ) => TOutputs | Promise<TOutputs>;
}